[IR] Convert profile metadata in createCallMatchingInvoke()
When an invoke instruction is converted to a call its
profile metadata is dropped because it has incompatible
format (see commit 2974ce1a9bc1).
This patch adds an attempt to convert profile data to
format of the call instruction. This used to work well
before the commit 64e2741d285e.

Reviewers: reames
Tags: #llvm
Differential Revision: https://reviews.llvm.org/D82071