[PowerPC] Fix crash for shuffle canonicalization with elt 0 from RHS
Commit dce5757187a6 assumed that shuffle vector canonicalization will
always ensure that the shuffle mask will be ordered so that element
zero comes from the LHS vector. However there is code out there for
which this is not the case. This patch simply removes that unsafe
assumption and makes the code work regardless of the source of the
first element.