[PowerPC] Remove undefs from splat input when changing shuffle mask
As of dce5757187a6d7dd9bf276ec0d908822aee52018, we have code that
changes shuffle masks so that we can put the shuffle in a canonical
form that can be matched to a single instruction. However, it
does not properly account for undef elements in the BUILD_VECTOR
that is the RHS splat so we can end up with undefs where they
shouldn't be. This patch converts the splat input with undefs to
one without.