[clangd] Fix the build with clang <3.9.
In clang <3.9 the `unique_ptr` constructor that is supposed to allow
for Derived to Base conversion does not work. Remove this if we drop
support for such configurations.

This is the same fix as in d7cd178032dec, and it updates the comments
to better reflect the actual issue. The same thing reproduces with
libc++ with older clangs.