Reland "[SCEVExpander] Add option to preserve LCSSA directly."
This reverts the revert commit 6e388e5b6021461edfb6f58a2c3d3ea5fdb214f1.

It includes a fix for Polly, which uses SCEVExpander on IR that is not
in LCSSA form. Set PreserveLCSSA = false in that case, to ensure we do
not introduce LCSSA phis where there were none before.