[AMDGPU] Fix crash when dag-combining bitcast
From the code after the 'break', they are processing 64bit scalar and
vector bitcast. So I think the break-condition should be (cond1 || cond2)
This means we only execute following code if (64bit and dest-is-vector).

Also remove a previous fix which is not needed with this new fix.
(introduced in: 3b7f0d35e49c59a056977d83d25b953f713ddd86)

Reviewed By: arsenm

Differential Revision: https://reviews.llvm.org/D85804